UIImage static function을 convinience initializer로 변경 및 파일명 변경
